[+] added random hero id in notification
