--- app/src/main/java/effective/office/marvelproject/di/NotificationModule.kt
@@ -1,80 +1,0 @@
-package effective.office.marvelproject.di
-
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.app.PendingIntent
-import android.app.TaskStackBuilder
-import android.content.Context
-import android.content.Intent
-import android.os.Build
-import androidx.core.app.NotificationCompat
-import androidx.core.app.NotificationManagerCompat
-import androidx.core.net.toUri
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dagger.hilt.components.SingletonComponent
-import effective.office.marvelproject.MainActivity
-import effective.office.marvelproject.R
-import effective.office.marvelproject.presentation.navigation.MY_URI
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object NotificationModule {
-
-    @Singleton
-    @Provides
-    fun provideNotificationBuilder(
-        @ApplicationContext context: Context
-    ): NotificationCompat.Builder {
-        val flag =
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M)
-                PendingIntent.FLAG_IMMUTABLE
-            else
-                0
-        val clickIntent = Intent(
-            Intent.ACTION_VIEW,
-            "$MY_URI/index=1011176".toUri(),
-            context,
-            MainActivity::class.java
-        )
-        val clickPendingIntent: PendingIntent = TaskStackBuilder.create(context).run {
-            addNextIntentWithParentStack(clickIntent)
-            getPendingIntent(1, flag)
-        }
-        return NotificationCompat.Builder(context, "Main Channel ID")
-            .setContentTitle("Welcome")
-            .setSmallIcon(R.drawable.ic_launcher_foreground)
-            .setAutoCancel(true)
-            .setContentText("YouTube Channel: Stevdza-San")
-            .setPriority(NotificationCompat.PRIORITY_DEFAULT)
-            .setVisibility(NotificationCompat.VISIBILITY_PRIVATE)
-            .setPublicVersion(
-                NotificationCompat.Builder(context, "Main Channel ID")
-                    .setContentTitle("Hidden")
-                    .setContentText("Unlock to see the message.")
-                    .build()
-            )
-            .setContentIntent(clickPendingIntent)
-    }
-
-    @Singleton
-    @Provides
-    fun provideNotificationManager(
-        @ApplicationContext context: Context
-    ): NotificationManagerCompat {
-        val notificationManager = NotificationManagerCompat.from(context)
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            val channel = NotificationChannel(
-                "Main Channel ID",
-                "Main Channel",
-                NotificationManager.IMPORTANCE_DEFAULT
-            )
-            notificationManager.createNotificationChannel(channel)
-        }
-        return notificationManager
-    }
-
-}
