+++ app/src/main/java/effective/office/marvelproject/presentation/navigation/NavGraph.kt
@@ -4,9 +4,7 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
 
-const val MY_URI = "https://stevdza-san.com"
-const val MY_ARG = 0
-
+const val MY_URI = "https://app.com"
 @Composable
 fun NavGraph() {
     val navController = rememberNavController()
