+++ app/src/main/java/effective/office/marvelproject/utils/Constant.kt
@@ -17,5 +17,8 @@ class Constant {
             val md = MessageDigest.getInstance("MD5")
             return BigInteger(1, md.digest(input.toByteArray())).toString(16).padStart(32, '0')
         }
+
+        const val CHANNEL_ID = "id_heroes"
+        const val CHANNEL_NAME = "hero_channel"
     }
 }
